Minor refactors for readability
